requirements.txt
opencv-python
pillow
python-chess
pyautogui